[Bronze II] Title: Hashing, Time: 0 ms, Memory: 2208 KB, Score: 50 point -BaekjoonHub

+++ 백준/Bronze/15829. Hashing/Hashing.cc
@@ -3,7 +3,7 @@
 
 using namespace std;
 
-unsigned long hashing(int l, const string& str) {
+unsigned long long hashing(int l, const string& str) {
     int result = 0;
     for (int i = 0; i < l; i++)
         result += ((int)str[i] - 96) * (int)pow(31, i);
